+++ src/constant/walletType.js
@@ -30,7 +30,7 @@ export const WALLET_CONNECT_TYPE = {
 /** not support transaction history */
 export const NET_CONFIG_NOT_SUPPORT_TX_HISTORY =[
   NET_CONFIG_TYPE.Unknown,
-  NET_CONFIG_TYPE.Devnet,
+  // NET_CONFIG_TYPE.Devnet,
 ]
 /** not support stake   */
 export const NET_CONFIG_NOT_SUPPORT_STAKING =[