+++ src/popup/pages/AboutUs/index.js
@@ -15,6 +15,7 @@ const AboutUs = ({}) => {
     privacy_policy_cn: "",
     changelog: "",
     followus: [],
+    graphql_api:""
   });
   const fecthBaseInfo = useCallback(async () => {
     let baseInfo = await getBaseInfo().catch((err) => err);
@@ -30,6 +31,7 @@ const AboutUs = ({}) => {
       privacy_policy_cn: baseInfo.privacy_policy_cn,
       changelog: baseInfo.changelog,
       followus: baseInfo.followus,
+      graphql_api:baseInfo.graphql_api,
     });
   }, []);
 
@@ -92,6 +94,10 @@ const AboutUs = ({}) => {
       {
         title: i18n.t("checkOnGithub"),
         link: baseAboutInfo.changelog,
+      },
+      {
+        title: i18n.t("serviceSupport"),
+        link: baseAboutInfo.graphql_api,
       },
     ];
     return { followList, linkInfoList };