+++ src/popup/pages/Wallet/index.js
@@ -7,7 +7,7 @@ import { Trans } from 'react-i18next';
 import { useDispatch, useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { cointypes } from '../../../../config';
-import { getBalance, getCurrencyPrice, getPendingTxList, getTransactionList } from "../../../background/api";
+import { getBalance, getCurrencyPrice, getGqlTxHistory, getPendingTxList, getTransactionList } from "../../../background/api";
 import { saveLocal } from '../../../background/localStorage';
 import { NET_WORK_CONFIG } from '../../../constant/storageKey';
 import { DAPP_DISCONNECT_SITE, DAPP_GET_CONNECT_STATUS, WALLET_GET_ALL_ACCOUNT } from '../../../constant/types';
@@ -401,38 +401,35 @@ const WalletDetail = () => {
 
   const [currentAccount, setCurrentAccount] = useState(accountInfo.currentAccount)
   const [historyList, setHistoryList] = useState(accountInfo.txList)
-  const [showHistoryStatus, setShowHistoryStatus] = useState(true)
+  const [showHistoryStatus, setShowHistoryStatus] = useState(()=>{
+    let status
+    let netType = netConfig?.currentConfig?.netType
+    if (getNetTypeNotSupportHistory(netType)) {
+      status = false
+    } else {
+      status = true
+    }
+    return status
+  })
   const [historyRefreshing, setHistoryRefreshing] = useState(false)
 
   const [loadingStatus, setLoadingStatus] = useState(false)
 
-  const isFirstRequest = useRef(true);
-
+  const isFirstRequest = useRef(historyList.length===0);
+ 
   const requestHistory = useCallback(async (silent = false, address = currentAccount.address) => {
     let netType = netConfig?.currentConfig?.netType
     if (!getNetTypeNotSupportHistory(netType)) {
-      if (!isFirstRequest.current && !silent) {
+      if (isFirstRequest.current && !silent) {
         setLoadingStatus(true)
       }
-      let currentAddress = currentAccount.address
-      let txList = getTransactionList(address)
       let pendingTxList = getPendingTxList(address)
-      await Promise.all([txList, pendingTxList]).then((data) => {
-        let txData = data[0]
-        let txAddress = txData.address
-        let txList = txData.txList
-
+      let gqlTxList = getGqlTxHistory(address)
+      await Promise.all([gqlTxList,pendingTxList]).then((data) => {
+        let newList = data[0]
         let txPendingData = data[1]
         let txPendingList = txPendingData.txList
-        let txPendingAddress = txPendingData.address
-        let newList = []
-        if (currentAddress === txPendingAddress) {
-          newList = [...txPendingList]
-        }
-        if (currentAddress === txAddress) {
-          newList = [...newList, ...txList]
-        }
-        dispatch(updateAccountTx(txList, txPendingList))
+        dispatch(updateAccountTx(newList,txPendingList))
       }).catch((err) => {
       }).finally(() => {
         setHistoryRefreshing(false)
@@ -447,7 +444,7 @@ const WalletDetail = () => {
 
   useEffect(() => {
     if (showHistoryStatus) {
-      requestHistory()
+      requestHistory(true)
     }
   }, [showHistoryStatus, requestHistory])
 
@@ -462,7 +459,7 @@ const WalletDetail = () => {
   }, [requestHistory])
 
 
-  const [showEmptyView, setShowEmptyView] = useState(false)
+  const [showNoBalanceView, setShowNoBalanceView] = useState(false)
 
 
   useEffect(() => {
@@ -494,25 +491,34 @@ const WalletDetail = () => {
 
   useEffect(() => {
     if (historyList.length === 0 && !loadingStatus && showHistoryStatus) {
-      setShowEmptyView(true)
+      setShowNoBalanceView(true)
     } else {
-      setShowEmptyView(false)
+      setShowNoBalanceView(false)
     }
   }, [historyList, showHistoryStatus, loadingStatus])
 
-  return (<div className={styles.walletDetail}>
-    {
-      showEmptyView ? <NoBalanceDetail /> :
-        <TxListView
-          history={historyList}
-          historyRefreshing={historyRefreshing}
-          onClickRefesh={onClickRefesh}
-          showEmpty={showEmptyView}
-          showHistoryStatus={showHistoryStatus}
-          loading={loadingStatus}
-        />
+  let childView =(<></>)
+  if(showNoBalanceView){
+    childView = <NoBalanceDetail /> 
+  }else{// 
+    if(loadingStatus){
+      childView = <LoadingView/>
+    }else {
+      if(historyList.length !== 0){
+        childView = <TxListView
+        history={historyList}
+        onClickRefesh={onClickRefesh}
+        historyRefreshing={historyRefreshing}
+        showHistoryStatus={showHistoryStatus}
+      />
+      }else{
+        childView =  <EmptyView/>
+      }
     }
 
+  }
+  return (<div className={styles.walletDetail}>
+    {childView}
     <Clock schemeEvent={() => { requestHistory(true) }} />
   </div>)
 }
@@ -523,8 +529,6 @@ const WalletDetail = () => {
  * @returns 
  */
 const TxListView = ({
-  showEmpty = false,
-  loading = false,
   history = [],
   showHistoryStatus = false,
   historyRefreshing = false,
@@ -533,35 +537,14 @@ const TxListView = ({
   const accountInfo = useSelector(state => state.accountInfo)
   const netConfig = useSelector(state => state.network)
 
-  const [showUnsupport, setShowUnsupport] = useState(getNetTypeNotSupportHistory(netConfig.currentConfig?.netType))
-
   const onGoExplorer = useCallback(() => {
     let currentConfig = netConfig.currentConfig
     let url = currentConfig.explorer + "/wallet/" + accountInfo.currentAccount.address
     openTab(url)
   }, [netConfig, accountInfo])
 
-
-  useEffect(() => {
-    setShowUnsupport(getNetTypeNotSupportHistory(netConfig.currentConfig?.netType))
-  }, [netConfig.currentConfig])
-
-  return (<div className={cls(styles.historyContainer, {
-    [styles.emptyList]: showEmpty,
-    [styles.loadingOuterContainer]: loading,
-    [styles.holderContainer]: !loading && !showEmpty,
-  })}>
-    <div className={styles.historyHead}>
-      <span className={styles.historyTitle}>{i18n.t("history")}</span>
-      {showHistoryStatus && <div className={cls(styles.refreshContainer,{
-        [styles.refresh]: historyRefreshing
-      })} onClick={onClickRefesh}>
-        <img src="/img/refresh.svg"  className={styles.refreshIcon}/>
-      </div>
-      }
-    </div>
- 
-    {!loading && (!showUnsupport ?
+  return (<div className={cls(styles.historyContainer,styles.holderContainer)}>
+    <HistoryHeader showHistoryStatus={showHistoryStatus} historyRefreshing={historyRefreshing} onClickRefesh={onClickRefesh}/>
       <div className={styles.listContainer}>
         {history.map((item, index) => {
           if (item.showExplorer) {
@@ -574,20 +557,46 @@ const TxListView = ({
           }
           return <TxItem txData={item} key={index} currentAccount={accountInfo.currentAccount} />
         })}
-      </div> :
+      </div> 
+  </div>)
+}
+const HistoryHeader=({historyRefreshing,onClickRefesh,showHistoryStatus})=>{
+  return (
+    <div className={styles.historyHead}>
+      <span className={styles.historyTitle}>{i18n.t("history")}</span>
+      {showHistoryStatus && <div className={cls(styles.refreshContainer,{
+        [styles.refresh]: historyRefreshing
+      })} onClick={onClickRefesh}>
+        <img src="/img/refresh.svg"  className={styles.refreshIcon}/>
+      </div>
+      }
+    </div>
+  )
+}
+const LoadingView=()=>{
+  return(
+    <div className={cls(styles.historyContainer,styles.loadingOuterContainer)}>
+      <HistoryHeader showHistoryStatus={false}/>
+      <div className={styles.loadingCon}>
+        <img className={styles.refreshLoading} src="/img/loading_purple.svg" />
+        <p className={styles.loadingContent}>
+          {i18n.t('loading') + "..."}
+        </p>
+      </div>
+    </div>
+  )
+}
+const EmptyView = ()=>{
+  return (<div className={styles.historyContainer}>
+    <div className={styles.historyHead}>
+      <span className={styles.historyTitle}>{i18n.t("history")}</span>
+    </div>
       <div className={styles.emptyContainer}>
         <img src="/img/icon_empty.svg" className={styles.emptyIcon} />
         <span className={styles.emptyContent}>
-          {i18n.t('notSupportTx')}
+          {i18n.t('unknownInfo')}
         </span>
-      </div>)}
-    {loading && <div className={styles.loadingCon}>
-      <img className={styles.refreshLoading} src="/img/loading_purple.svg" />
-      <p className={styles.loadingContent}>
-        {i18n.t('loading') + "..."}
-      </p>
-    </div>}
-
+      </div>
   </div>)
 }
 const NoBalanceDetail = () => {
@@ -624,37 +633,37 @@ const TxItem = ({ txData, currentAccount }) => {
     let statusIcon, showAddress, timeInfo, amount, statusText, statusStyle = ''
 
 
-    if (txData.type.toLowerCase() === "payment") {
-      isReceive = txData.receiver.toLowerCase() === currentAccount.address.toLowerCase()
+    if (txData.kind.toLowerCase() === "payment") {
+      isReceive = txData.to.toLowerCase() === currentAccount.address.toLowerCase()
       statusIcon = isReceive ? "/img/tx_receive.svg" : "/img/tx_send.svg"
     } else {
       statusIcon = "/img/tx_pending.svg"
     }
 
-    showAddress = addressSlice(isReceive ? txData.sender : txData.receiver, 8)
-    showAddress = !showAddress ? txData.type.toUpperCase() : showAddress
+    showAddress = addressSlice(isReceive ? txData.from : txData.to, 8)
+    showAddress = !showAddress ? txData.kind.toUpperCase() : showAddress
 
-    timeInfo = txData.status === TX_STATUS.PENDING ? "Nonce " + txData.nonce : getShowTime(txData.time)
+    timeInfo = txData.status === TX_STATUS.PENDING ? "Nonce " + txData.nonce : getShowTime(txData.dateTime)
 
 
     amount = amountDecimals(txData.amount, cointypes.decimals)
     amount = getDisplayAmount(amount, 2)
     amount = isReceive ? "+" + amount : "-" + amount
 
-    statusText = i18n.t(txData.status && txData.status.toUpperCase())
-
     let showPendTx = false
-    switch (txData.status) {
-      case TX_STATUS.SUCCESS:
-        statusStyle = styles.itemStatus_Success
-        break;
-      case TX_STATUS.FAILED:
+
+    if(txData.status === TX_STATUS.PENDING){
+      statusStyle = styles.itemStatus_Pending
+      showPendTx = true
+      statusText = i18n.t(txData.status && txData.status.toUpperCase()) 
+    }else{
+      if(txData.failureReason){
         statusStyle = styles.itemStatus_Failed
-        break;
-      case TX_STATUS.PENDING:
-        statusStyle = styles.itemStatus_Pending
-        showPendTx = true
-        break;
+        statusText = i18n.t("FAILED")
+      }else{
+        statusStyle = styles.itemStatus_Success
+        statusText = i18n.t("APPLIED")
+      }
     }
 
     return {